+++ backend/src/main/java/com/project/whalearc/trade/service/LimitOrderScheduler.java
@@ -10,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -77,7 +78,7 @@ public class LimitOrderScheduler {
             if (marketPrice == null) continue;
 
             try {
-                if (orderService.tryExecuteLimitOrder(order, marketPrice)) {
+                if (orderService.tryExecuteLimitOrder(order, BigDecimal.valueOf(marketPrice))) {
                     executed++;
                 }
             } catch (Exception e) {
